Use offensive instead of separate algorithm
The issue here was the profane? algorithm flagged words
that were not actually offensive.
Specifically, it flagged any string more than 2 characters long
that contained `as` as profane.
Whereas, the same string did not contain any offensive characters.

These two methods should agree on their logic of which words are
offensive and profane.